refactor: remove GenesisInfo struct
