+++ examples/transfer_from_multisig.rs
@@ -14,8 +14,7 @@ use ckb_sdk::{
     },
     tx_builder::{transfer::CapacityTransferBuilder, unlock_tx, CapacityBalancer, TxBuilder},
     unlock::{MultisigConfig, ScriptUnlocker, SecpMultisigScriptSigner, SecpMultisigUnlocker},
-    Address, AddressPayload, GenesisInfo, HumanCapacity, NetworkType, ScriptGroup, ScriptId,
-    SECP256K1,
+    Address, AddressPayload, HumanCapacity, NetworkType, ScriptGroup, ScriptId, SECP256K1,
 };
 use ckb_types::{
     bytes::Bytes,
@@ -244,8 +243,7 @@ fn build_transfer_tx(
     let mut ckb_client = CkbRpcClient::new(args.ckb_rpc.as_str());
     let cell_dep_resolver = {
         let genesis_block = ckb_client.get_block_by_number(0.into())?.unwrap();
-        let info = GenesisInfo::from_block(&BlockView::from(genesis_block))?;
-        DefaultCellDepResolver::new(&info)
+        DefaultCellDepResolver::from_genesis(&BlockView::from(genesis_block))?
     };
     let header_dep_resolver = DefaultHeaderDepResolver::new(args.ckb_rpc.as_str());
     let mut cell_collector =
