+++ src/lib.rs
@@ -1,5 +1,3 @@
-mod chain;
-
 pub mod bip32;
 pub mod constants;
 pub mod pubsub;
@@ -17,7 +15,6 @@ pub mod test_util;
 #[cfg(test)]
 mod tests;
 
-pub use chain::{calc_max_mature_number, GenesisInfo, ParseGenesisInfoError};
 pub use rpc::{CkbRpcClient, IndexerRpcClient, RpcError};
 pub use types::{
     Address, AddressPayload, AddressType, CodeHashIndex, HumanCapacity, NetworkType, OldAddress,
